split mapservice test files
